Migrate to python3.8 using the most recent archive
The migration to recent archives (2019 was used until now), required to migrate
to an instance random generator instead of using the numpy global one because of
dask using threads. The combination of the global numpy generator, which is not
thread-safe, with dask threads resulted to non-reproducible results, changing at
every run even though seed was fixed. Unfortunately, it was not possible to
reproduce the same results using the global regenerator and the legacy RandomState.
The reason is unknown but it could be related to a bug in diameter-synthesis, which
was loading a seed from the parameters file, resetting the seed set by the worker.

The latest random generator (default_rng, PCG64) will be used from now on, which provides
reproducible results and it is not affected by the parallel processing, because each
worker creates and uses an instance of the generator that is not global.

+++ archngv/spatial/bounding_box.py
@@ -26,7 +26,7 @@ class BoundingBox:
 
     def __init__(self, min_coordinates, max_coordinates):
         self._bb = np.array((min_coordinates,
-                             max_coordinates), dtype=np.float)
+                             max_coordinates), dtype=np.float32)
 
     def __eq__(self, other):
         """ Equality of bboxes"""
